+++ server/aws-lsp-codewhisperer/src/language-server/amazonQServiceManager/BaseAmazonQServiceManager.ts
@@ -1,0 +1,7 @@
+import { DidChangeConfigurationParams } from '@aws/language-server-runtimes/protocol'
+import { CodeWhispererServiceBase } from '../codeWhispererService'
+
+export interface BaseAmazonQServiceManager {
+    handleDidChangeConfiguration: (ideCategory?: string, params?: DidChangeConfigurationParams) => Promise<void>
+    getCodewhispererService: () => CodeWhispererServiceBase
+}