+++ core/aws-lsp-core/CHANGELOG.md
@@ -1,3 +1,12 @@
+# Changelog
+
+## [0.0.2](https://github.com/aws/language-servers/compare/lsp-core/v0.0.1...lsp-core/v0.0.2) (2025-03-18)
+
+
+### Features
+
+* **identity:** device code support ([#823](https://github.com/aws/language-servers/issues/823)) ([6d5368e](https://github.com/aws/language-servers/commit/6d5368e33a36a3003dc04e9c429b63edda6989de))
+
 [0.0.1] - 2024-06-07
 
 - Intial release containing helper classes and functions that are used in `@aws/lsp-yaml`, `@aws/lsp-json` and other packages