+++ server/aws-lsp-json/CHANGELOG.md
@@ -1,3 +1,26 @@
+# Changelog
+
+## [0.1.2](https://github.com/aws/language-servers/compare/lsp-json/v0.1.1...lsp-json/v0.1.2) (2025-03-26)
+
+
+### Bug Fixes
+
+* update @aws/language-server-runtimes to 0.2.48 ([e1f620a](https://github.com/aws/language-servers/commit/e1f620ac2b59b4f61daff842a9f29ded1b8fa04e))
+
+## [0.1.1](https://github.com/aws/language-servers/compare/lsp-json/v0.1.0...lsp-json/v0.1.1) (2025-03-18)
+
+
+### Features
+
+* **chat-client:** handle 'openTab' requests ([#817](https://github.com/aws/language-servers/issues/817)) ([fdd0b87](https://github.com/aws/language-servers/commit/fdd0b87ad2d2c9a540d2594bb9243cad01b5887a))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @aws/lsp-core bumped from ^0.0.1 to ^0.0.2
+
 [0.0.1] - 2024-06-07
 
 - Intial release of the JSON LSP Server